% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_raster_summary.R
\name{aggregate_signals}
\alias{aggregate_signals}
\title{aggregate_signals}
\usage{
aggregate_signals(
  profile_dt,
  agg_FUN = max,
  y_ = "y",
  yout_ = "ynorm",
  xmin = -Inf,
  xmax = Inf,
  by_ = c("tall_var", "wide_var")
)
}
\arguments{
\item{profile_dt}{Tidy data.table of profile information. As returned by seqsetvis::ssvFetchBam.}

\item{agg_FUN}{A function appled to all y_ values in xmin to xmax range. Must accept single numeric vector and return 1 value.}

\item{y_}{Variable name containing profile scores.}

\item{yout_}{Variable name in output containing aggregate profile score information.}

\item{xmin}{The min range of x-values to apply agg_FUN to}

\item{xmax}{The max range of x-values to apply agg_FUN to}

\item{by_}{Character vector of variable names to apply agg_FUN by.}
}
\value{
A data.table summarizing every profile with a single value. Suitable for input to chiptsne:::plot_binned_aggregates.
}
\description{
aggregate_signals is retrieves a single representative value for a genomic
region per sample by default this is the maximum anywhere in the region but
this can be overriden using xmin/xmax and agg_FUN
}
\examples{
aggregate_signals(profile_dt)
aggregate_signals(profile_dt, xmin = -.2, xmax = .2, agg_FUN = mean)
}
