% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{bin_values}
\alias{bin_values}
\title{bin_values}
\usage{
bin_values(x, n_bins, xrng = range(x))
}
\arguments{
\item{x}{Values to assign to bins}

\item{n_bins}{Number of bins to assign values to}

\item{xrng}{Optional numeric of length 2. Defines the domain to define bins
for. Defaults to range(x).}
}
\value{
bin assignments parall to x.
}
\description{
bin_values
}
\examples{
bin_values(0:10, 3)
}
