% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{bin_values_centers}
\alias{bin_values_centers}
\title{bin_values_centers}
\usage{
bin_values_centers(n_bins, rng)
}
\arguments{
\item{n_bins}{Number of bins to assign values to}

\item{rng}{Numeric to define bins
for. Defaults to range(x).}
}
\value{
numeric of length n_bins containing centers of bins.
}
\description{
bin_values_centers
}
\examples{
bin_values_centers(3, 0:10)
bin_values_centers(3, range(0:10))
}
