% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPtSNE.R
\name{ChIPtSNE}
\alias{ChIPtSNE}
\title{ChIPtSNE}
\usage{
ChIPtSNE(
  features_config = NULL,
  signal_config = NULL,
  out_dir = getwd(),
  bfc = NULL,
  perplexity = 50,
  n_glyphs_x = 8,
  n_glyphs_y = n_glyphs_x,
  n_heatmap_pixels_x = 25,
  n_heatmap_pixels_y = n_heatmap_pixels_x,
  dimreduce_method = valid_dimreduce_methods$tsne
)
}
\arguments{
\item{features_config}{Controls features configuration.  May be a:
QcConfigFeatures object, path to a file defining configuration via
QcConfigFeatures.parse, features files to define via
QcConfigFeatures.files, or a data.frame to pass to QcConfigFeatures.}

\item{signal_config}{Controls signal configuration.  May be a: QcConfigSignal
object, path to a file defining configuration via QcConfigSignal.parse,
features files to define via QcConfigSignal.files, or a data.frame to pass
to QcConfigSignal.}

\item{out_dir}{NYI}

\item{bfc}{BiocFileCache object to use for caching. If NULL, default
new_cache() will be used.}

\item{perplexity}{passed to Rtsne::Rtsne()}

\item{n_glyphs_x}{number of glyphs across x-axis to summarize profiles in t-SNE space. Default is 8.}

\item{n_glyphs_y}{number of glyphs across y-axis. Default is the same as n_glyphs_x.}

\item{n_heatmap_pixels_x}{number of pixels to use in heatmap across x-axis in t-SNE space. Default is 25.}

\item{n_heatmap_pixels_y}{number of pixels to use in heatmap across y-axis in t-SNE space. Default is same as n_heatmap_pixels_x.}
}
\value{
a valid ChIPtSNE object.
}
\description{
ChIPtSNE
}
\examples{
library(chiptsne)
library(ssvQC)
query_bed_file = system.file(package = "chiptsne", "extdata/query_gr.bed")
query_gr = rtracklayer::import.bed(query_bed_file)
bam_files = dir(system.file(package = "chiptsne", "extdata"), pattern = "bam$", full.names = TRUE)
qc_config_features = QcConfigFeatures.GRanges(query_gr)
qc_config_signal = QcConfigSignal.files(bam_files)
qc_config_signal$flip_signal_mode = SQC_FLIP_SIGNAL_MODES$high_on_left
qc_config_signal$center_signal_at_max = TRUE
ct = ChIPtSNE(qc_config_features, qc_config_signal)
ct = ChIPtSNE.runTSNE(ct)

ct$n_glyphs_x = 7
ct$n_glyphs_y = 5

ct$n_heatmap_pixels_x = 5
ct$n_heatmap_pixels_y = 9
ct.replot = ssvQC.plotSignal(ct)

ct$plots$TSNE$regional_glyphs$query_features$All_signal
ct.replot$plots$TSNE$regional_glyphs$query_features$All_signal

ct$plots$TSNE$regional_heatmap$query_features$All_signal
ct.replot$plots$TSNE$regional_heatmap$query_features$All_signal

}
