% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIPtSNE.R
\name{ChIPtSNE.runTSNE}
\alias{ChIPtSNE.runTSNE}
\title{ChIPtSNE.runTSNE}
\usage{
ChIPtSNE.runTSNE(sts)
}
\arguments{
\item{sts}{A valid ChIPtSNE object, needs ssvQC.prepSignal to be called.}
}
\value{
A valid ChIPtSNE object with t-SNE called and ready for plots.
}
\description{
ChIPtSNE.runTSNE
}
\examples{
library(chiptsne)
library(ssvQC)
query_bed_file = system.file(package = "chiptsne", "extdata/query_gr.bed")
query_gr = rtracklayer::import.bed(query_bed_file)
bam_files = dir(system.file(package = "chiptsne", "extdata"), pattern = "bam$", full.names = TRUE)
qc_config_features = QcConfigFeatures.GRanges(query_gr)
qc_config_signal = QcConfigSignal.files(bam_files)
qc_config_signal$flip_signal_mode = SQC_FLIP_SIGNAL_MODES$high_on_left
qc_config_signal$center_signal_at_max = TRUE
sts = ChIPtSNE(qc_config_features, qc_config_signal)
sts = ChIPtSNE.runTSNE(sts)
}
