% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal_TSNE.R
\docType{class}
\name{ClusteredSignal_TSNE-class}
\alias{ClusteredSignal_TSNE-class}
\title{ClusteredSignal_TSNE}
\description{
adds support for xy coordinate (tsne result) data to ClusteredSignal class.
}
\section{Slots}{

\describe{
\item{\code{xy_data}}{data.table.}
}}

