% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal_TSNE.R
\name{ClusteredSignal_TSNE.null}
\alias{ClusteredSignal_TSNE.null}
\title{ClusteredSignal_TSNE.null}
\usage{
ClusteredSignal_TSNE.null()
}
\value{
A "null" ClusteredSignal_TSNE object containing no information.
}
\description{
ClusteredSignal_TSNE.null
}
\examples{
ClusteredSignal_TSNE.null()
}
