% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusteredSignal_TSNE.R
\name{ClusteredSignal_TSNE}
\alias{ClusteredSignal_TSNE}
\title{Title}
\usage{
ClusteredSignal_TSNE(
  signal_profile_dt,
  query_gr,
  manual_assigned = list(),
  nclust = 6,
  signal_var = "y",
  signal_var.within = "y",
  facet_var = "name_split",
  extra_var = character()
)
}
\arguments{
\item{extra_var}{}
}
\description{
Title
}
