% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cluster.R
\name{combine_mostsimilar}
\alias{combine_mostsimilar}
\title{combine_mostsimilar}
\usage{
combine_mostsimilar(
  p_dt,
  profile_dt,
  n_times = 1,
  min_dist = Inf,
  cluster_ = "cluster_id",
  new_cluster_ = cluster_
)
}
\arguments{
\item{min_dist}{}
}
\description{
combine_mostsimilar
}
