% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctPlots.R
\name{ctClusterPoints}
\alias{ctClusterPoints}
\title{ctClusterPoints}
\usage{
ctClusterPoints(sts, feature_name = NULL, signal_name = NULL, n_clust = 6)
}
\arguments{
\item{sts}{A ssvTSNE object with ssvQC.prepSignal already called.}

\item{n_clust}{number of nearest neighbor clusters to calculate.  Reached to calling too many clusters and iteratively combining most similar clusters.}
}
\value{
A data.table containing id and cluster_id information. Suitable for ctPlotPointsAnnotation.
}
\description{
ctClusterPoints
}
\examples{
data(sts.test)
clust_dt =ctClusterPoints(sts, n_clust = 3)

ctPlotPointsAnnotation(sts, meta_data = clust_dt, anno_var = "cluster_id")
}
