% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctPlots.R
\name{ctPlotPoints}
\alias{ctPlotPoints}
\title{ctPlotPoints}
\usage{
ctPlotPoints(
  sts,
  xmin = -Inf,
  xmax = Inf,
  profile_value = ssvQC:::val2var[sts$signal_config$plot_value],
  profile_value_label = ssvQC:::val2lab[sts$signal_config$plot_value],
  bg_color = "gray20",
  agg_FUN = max
)
}
\arguments{
\item{agg_FUN}{}
}
\description{
ctPlotPoints
}
\examples{
data(sts.test)
ctPlotPoints(sts)

}
