% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{fetch_tsne_mat}
\alias{fetch_tsne_mat}
\title{Title}
\usage{
fetch_tsne_mat(qdt, qgr, qwin = 50, qmet = "summary", cap_value = 20,
  high_on_right = TRUE, bfc = BiocFileCache::BiocFileCache(),
  n_cores = getOption("mc.cores", 1), rname = digest::digest(list(qgr,
  qdt[, 1:3], qwin, qmet, cap_value, high_on_right)),
  force_overwrite = FALSE)
}
\arguments{
\item{force_overwrite}{}
}
\description{
Title
}
