% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{make_img_plots}
\alias{make_img_plots}
\title{Title}
\usage{
make_img_plots(img_results, qcell = NULL, xrng = c(-0.5, 0.5),
  yrng = c(-0.5, 0.5), N_floor = 0, N_ceiling = NULL,
  min_size = 0.3, as_facet = TRUE)
}
\arguments{
\item{as_facet}{}
}
\description{
Title
}
