% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{make_img_plots_facet}
\alias{make_img_plots_facet}
\title{Title}
\usage{
make_img_plots_facet(img_results, qcell = NULL, xrng = c(-0.5, 0.5),
  yrng = c(-0.5, 0.5), N_floor = 0, N_ceiling = NULL,
  min_size = 0.3, facet_by = "cell")
}
\arguments{
\item{facet_by}{}
}
\description{
Title
}
