% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{make_tsne_img}
\alias{make_tsne_img}
\title{Title}
\usage{
make_tsne_img(profiles_dt, position_dt, n_points,
  xrng = range(position_dt$tx), yrng = range(position_dt$ty),
  rname = digest::digest(list(profiles_dt, position_dt, n_points,
  apply_norm, ylim, line_colors, facet_by)),
  odir = file.path("tsne_images/", rname), force_rewrite = FALSE,
  apply_norm = TRUE, ylim = c(0, 1), facet_by = NULL, ma_size = 2,
  n_cores = getOption("mc.cores", 1), line_colors = c(H3K4me3 =
  "forestgreen", H3K27me3 = "firebrick1"))
}
\arguments{
\item{line_colors}{}
}
\description{
Title
}
