% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{nn_clust}
\alias{nn_clust}
\title{nn_clust}
\usage{
nn_clust(
  tsne_res,
  nsamp = Inf,
  nn = 100,
  tall_var = "tall_var",
  id_var = "id",
  show_plot = FALSE,
  return_plot = FALSE
)
}
\arguments{
\item{tsne_res}{data.table with tsne results. (needs tx, ty, tall_var, id)}

\item{nsamp}{Downsample profile values down to this value. Default of Inf skips downsampling.}

\item{nn}{Number of neighbors to use for clustering.  Default of 100.}

\item{show_plot}{If TRUE, print plot to graphics device. Default is FALSE.}

\item{return_plot}{If TRUE, return plot in list with cluster results. Default is FALSE.}
}
\description{
performs nearest neighbor clustering on tsne coordinates.
}
\examples{
data("profile_dt")
data("tsne_dt")
setalloccol(tsne_dt)
clust_res = nn_clust(tsne_dt, nn = 5, return_plot = TRUE)
clust_res$plot
clust_res$data

#just data by default
nn_clust(tsne_dt, nn = 5)
}
