% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{plot_flat}
\alias{plot_flat}
\title{plot_flat}
\usage{
plot_flat(p, dpi = 150, width = dev.size()[1],
  height = dev.size()[2], panel_only = FALSE)
}
\arguments{
\item{p}{a ggplot object to render}

\item{dpi}{dots per inch for temp png, default is 150}

\item{width}{width in inches for temp png, default it 8}

\item{height}{height in inches for temp png, default is 8}
}
\value{
a ggplot object containing flat image of original
}
\description{
plot to a temporary png and redraw as a flat image
}
\details{
useful when svg would be too expensive due to high number of items
}
