% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_path.R
\name{plot_outline}
\alias{plot_outline}
\title{Title}
\usage{
plot_outline(tsne_dt, qgr, qcells, p = NULL, id_to_plot = NULL,
  xrng = c(-0.5, 0.5), yrng = c(-0.5, 0.5), bg_color = "gray",
  line_color_mapping = "black", fill_color_mapping = "gray",
  label_type = c("text", "label", "none")[3], bg_points = 5000,
  arrow_FUN = NULL)
}
\arguments{
\item{arrow_FUN}{}
}
\description{
Title
}
