% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_path.R
\name{plot_path}
\alias{plot_path}
\title{Title}
\usage{
plot_path(tsne_dt, qgr, qcells, tss_ids, grp_var = c("id", "grp")[1],
  line_type = c("curve", "spline", "straight")[2],
  label_type = c("text", "label", "none")[2], bg_points = 5000,
  arrow_FUN = NULL)
}
\arguments{
\item{label_type}{}
}
\description{
Title
}
