% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{plot_tsne_img}
\alias{plot_tsne_img}
\title{Title}
\usage{
plot_tsne_img(image_dt, n_points, p = NULL, xrng = c(-0.5, 0.5),
  yrng = c(-0.5, 0.5), N_floor = 0, N_ceiling = NULL,
  min_size = 0.3, show_plot = FALSE)
}
\arguments{
\item{show_plot}{}
}
\description{
Title
}
