% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{plot_velocity_arrows_binned}
\alias{plot_velocity_arrows_binned}
\title{Title}
\usage{
plot_velocity_arrows_binned(tsne_res, cell_a, cell_b, n_points, p = NULL,
  min_N = 0, id_to_plot = NULL)
}
\arguments{
\item{id_to_plot}{}
}
\description{
Title
}
