% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{plot_velocity_arrows_selected}
\alias{plot_velocity_arrows_selected}
\title{Title}
\usage{
plot_velocity_arrows_selected(tsne_res, qgr, qcells, tss_ids,
  grp_var = c("id", "grp")[1], line_type = c("curve", "spline",
  "straight")[2], label_type = c("text", "label", "none")[2])
}
\arguments{
\item{label_type}{}
}
\description{
Title
}
