% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{plot_velocity_bins}
\alias{plot_velocity_bins}
\title{Title}
\usage{
plot_velocity_bins(velocity_dt, bins = 36, bin_FUN = list(length, sum,
  mean)[[1]])
}
\arguments{
\item{velocity_dt}{}
}
\description{
Title
}
