% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{prep_images}
\alias{prep_images}
\title{prep_images}
\usage{
prep_images(profile_dt, position_dt, x_points, y_points = x_points,
  xrng = range(position_dt$tx), yrng = range(position_dt$ty),
  rname = digest::digest(list(profile_dt, position_dt, x_points,
  y_points, apply_norm, ylim, line_color_mapping, n_splines, ma_size,
  facet_by)), odir = file.path(tempdir(), rname),
  force_rewrite = FALSE, apply_norm = TRUE, ylim = c(0, 1),
  facet_by = NULL, ma_size = 2, n_splines = 10,
  n_cores = getOption("mc.cores", 1), line_color_mapping = NULL,
  vertical_facet_mapping = NULL)
}
\arguments{
\item{profile_dt}{a tidy data.table for profile data as retrieved by
stsFetchTsneInput.  Expected variable names are id, cell, mark, x, and y.}

\item{position_dt}{a tidy data.table containing t-sne embedding.  Expected
variable names are tx, ty, id, and cell.}

\item{x_points}{numeric.  number of grid points to use in x dimension.}

\item{y_points}{numeric.  number of grid points to use in y dimension.
Defaults to same value as x_points.}

\item{xrng}{view domain in x dimension, default is range of position_dt$tx.}

\item{yrng}{view domain in y dimension, default is range of position_dt$ty.}

\item{rname}{prefix for image files.  existing image files are used if
present.}

\item{odir}{output directory for image files.}

\item{force_rewrite}{if TRUE, images are overwritten even if they exist.}

\item{apply_norm}{if TRUE, y values are trimmed to 95th percentile and
transformed ot domain of [0,1]. Default is TRUE.}

\item{ylim}{y-limits of regional summary plots.  Default of c(0, 1) is
compatible with apply_norm = TRUE.}

\item{facet_by}{character. varaible name to facet profile_dt by when
constructing images. The only valid non-null value with seqtsne functions
is "cell".}

\item{ma_size}{moving average size when smoothing profiles.}

\item{n_splines}{number of points to interpolate with splines.}

\item{n_cores}{number of cores to use when writing images.  Default is
value of mc.cores option if set or 1.}

\item{line_color_mapping}{named vector of line color.  Names correspond to values
of profile_dt 'mark' variable and values are colors.}

\item{vertical_facet_mapping}{named vector of vertical facet for data}
}
\value{
data.table with variables
}
\description{
Prepares images summarizing profiles in t-sne regions and returns ggimage
compatible data.frame for plotting.
}
\examples{
data("profile_dt")
data("tsne_dt")
img_res = prep_images(profile_dt, tsne_dt, 4)
#zoom on top-right quadrant
img_res.zoom = prep_images(profile_dt, tsne_dt, 4, xrng = c(0, .5), yrng = c(0, .5))
#use results with plot_tsne_img() to make plots
}
