% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{prep_profile_dt}
\alias{prep_profile_dt}
\title{applies normalizations and transformations to prof_dt, a tidy data.table of
profiles.}
\usage{
prep_profile_dt(prof_dt, norm_dt, qgr, cap_value, high_on_right = TRUE)
}
\arguments{
\item{high_on_right}{}
}
\description{
applies normalizations and transformations to prof_dt, a tidy data.table of
profiles.
}
