% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{prep_query_gr}
\alias{prep_query_gr}
\title{prepares query_gr GRanges for using in fetch functions}
\usage{
prep_query_gr(query_gr, region_size = NULL, id_prefix = "region")
}
\arguments{
\item{query_gr}{GRanges of regions to be fetched}

\item{region_size}{numeric fixed width to apply to query_gr. if NULL, width
is not enforced.}

\item{id_prefix}{character prefix to use for ids. 'region' is default and
will yield ids such as region_1, region_2, region_3 ...}
}
\value{
GRanges ready for fetching
}
\description{
fixed widths are recommended but not required
}
\details{
metadata column id must be set and all ids must be unique
}
