% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{prep_tsne_img}
\alias{prep_tsne_img}
\title{Title}
\usage{
prep_tsne_img(simg_dt, n_points, xrng, yrng, N_floor = 0,
  N_ceiling = NULL, min_size = 0.3)
}
\arguments{
\item{min_size}{}
}
\description{
Title
}
