% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{prep_velocity}
\alias{prep_velocity}
\title{prep_velocity}
\usage{
prep_velocity(tsne_dt, cell_a, cell_b, id_to_plot = NULL,
  max_plotted = 500, delta.min = 0, delta.max = Inf, angle.min = 0,
  angle.max = 360, drop_backgroud = FALSE)
}
\arguments{
\item{tsne_dt}{tidy data.table of tsne results.  variables names must
include: tx, ty, id, and cell}

\item{cell_a}{an item in tsne_dt$cell. the origin for calculating
angle/distance.}

\item{cell_b}{an item in tsne_dt$cell. the destination for calculating
angle/distance.}

\item{id_to_plot}{character.  ids in tsne_dt$id. Default of NULL causes all
ids to be used.}

\item{max_plotted}{the maximum number of ids to plot.}

\item{delta.min}{numeric. ids below this delta marked as background.}

\item{delta.max}{numeric. ids above this delta marked as background.}

\item{angle.min}{numeric. ids with angle below this marked as background.}

\item{angle.max}{numeric. ids with angle above this marked as background.}

\item{drop_backgroud}{if TRUE, all background velocities are dropped. Default
is FALSE.}
}
\value{
a tidy data.table of velocity measurements from cell_a to cell_b
}
\description{
prep_velocity
}
\examples{
data("tsne_dt")
vel_dt = prep_velocity(tsne_dt,
    unique(tsne_dt$cell)[2], unique(tsne_dt$cell)[3])
vel_dt
}
