% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tsne.R
\docType{data}
\name{query_gr}
\alias{query_gr}
\title{GRanges region annotation of hESC bivalent sites.}
\format{
A GRanges with 30 rows and mcols:
\describe{
  \item{group}{a bivalency category from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5354816/}{PMC5354816}}
  \item{id}{an arbitrary unique identifier}
}
}
\source{
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17312}{GSE17312}
}
\description{
\code{\link{hESC_data}}
}
\keyword{datasets}
