% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{rescale_capped}
\alias{rescale_capped}
\title{wraps scales::rescale to enforce range of \code{to} on output}
\usage{
rescale_capped(x, to = c(0, 1), from = range(x, na.rm = TRUE, finite =
  TRUE))
}
\arguments{
\item{x}{continuous vector of values to manipulate.}

\item{to}{output range (numeric vector of length two)}

\item{from}{input range (vector of length two). If not given, is calculated from the range of x}
}
\value{
x rescaled from \code{from} domain to \code{to} domain, within bounds of \code{to}
}
\description{
wraps scales::rescale to enforce range of \code{to} on output
}
\examples{
#behaves identically to scales::rescale when x is within 'from' domain
rescale_capped(0:10, to = c(0, 1), c(0, 10))
scales::rescale(0:10, to = c(0, 1), c(0, 10))
#when x exceeds 'from' domain, results are still within 'to' domain
rescale_capped(0:10, to = c(0,1), c(0,5))
#not true for scales::rescale
scales::rescale(0:10, to = c(0,1), c(0,5))
}
