% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{run_tsne}
\alias{run_tsne}
\title{Title}
\usage{
run_tsne(tsne_mat, perplexity = 100, n_cores = getOption("mc.cores",
  1), high_topright = TRUE, norm1 = TRUE,
  bfc = BiocFileCache::BiocFileCache(),
  rname = digest::digest(list(tsne_mat[sample(1:nrow(tsne_mat), 20),
  sample(1:ncol(tsne_mat), 20)], perplexity)), force_overwrite = FALSE)
}
\arguments{
\item{force_overwrite}{}
}
\description{
Title
}
