% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{runApp.chiptsne}
\alias{runApp.chiptsne}
\title{runApp.chiptsne}
\usage{
runApp.chiptsne(ct)
}
\arguments{
\item{ct}{}
}
\value{
Runs a shiny app
}
\description{
runApp.chiptsne
}
\examples{
library(IKdata)
ct = readRDS("presentation_ChIPtSNE_objects/4_combinatorial.Rds")
runApp.chiptsne(ct)
}
