% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{sampleCap}
\alias{sampleCap}
\title{sampleCap}
\usage{
sampleCap(x, n = 500)
}
\arguments{
\item{x}{vector to sample from}

\item{n}{number of items to sample.  automatically reduced to length of x if
too high.}
}
\value{
x sampled down to n items.
}
\description{
Handy wrapper to sample() that avoid having to check against exceeding size
of x.
}
\examples{
x = seq(10)
sampleCap(x, 5)
#avoid having to check number of items being sampled
sampleCap(x, 15)
}
