% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvTSNE.R
\name{ssvTSNE.runTSNE}
\alias{ssvTSNE.runTSNE}
\title{ssvTSNE.runTSNE}
\usage{
ssvTSNE.runTSNE(sts)
}
\arguments{
\item{sts}{A valid ssvTSNE object, needs ssvQC.prepSignal to be called.}
}
\value{
A valid ssvTSNE object with t-SNE called and ready for plots.
}
\description{
ssvTSNE.runTSNE
}
\examples{
data(sts.test)
ssvTSNE.runTSNE(sts)
}
