% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{stsFetchTsneInput}
\alias{stsFetchTsneInput}
\title{Title}
\usage{
stsFetchTsneInput(qdt, qgr, region_size = NULL, file_format = NULL,
  qwin = 50, qmet = "summary", cap_value = 20, agg_FUN = NULL,
  high_on_right = TRUE, bfc = BiocFileCache::BiocFileCache(),
  n_cores = getOption("mc.cores", 1), rname = digest::digest(list(qgr,
  qdt[, 1:3], qwin, qmet)), force_overwrite = FALSE)
}
\arguments{
\item{force_overwrite}{}
}
\description{
Title
}
