% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cluster.R
\name{stsPlotClusterProfiles}
\alias{stsPlotClusterProfiles}
\title{Title}
\usage{
stsPlotClusterProfiles(
  profile_dt,
  cluster_dt = profile_dt,
  cluster_ = "cluster_id",
  w = 0.05,
  h = 0.05,
  x_var = "x",
  y_var = "y",
  id_var = "id",
  wide_var = "name",
  tall_var = "tall_none"
)
}
\arguments{
\item{h}{}
}
\description{
Title
}
