% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{stsRunTsne}
\alias{stsRunTsne}
\title{Title}
\usage{
stsRunTsne(tsne_mat, perplexity = 100, n_cores = getOption("mc.cores",
  1), high_topright = TRUE, norm1 = TRUE,
  bfc = BiocFileCache::BiocFileCache(), rname = NULL,
  force_overwrite = FALSE)
}
\arguments{
\item{force_overwrite}{}
}
\description{
Title
}
