% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_raster_summary.R
\name{aggregate_signals}
\alias{aggregate_signals}
\title{aggregate_signals}
\usage{
aggregate_signals(profile_dt, agg_FUN = max, xmin = -Inf, xmax = Inf,
  by_ = c("tall_var", "wide_var"))
}
\arguments{
\item{by_}{}
}
\description{
aggregate_signals
}
\examples{
aggregate_signals(profile_dt)
aggregate_signals(profile_dt, xmin = -.2, xmax = .2, agg_FUN = mean)
}
