% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{bin_values_centers}
\alias{bin_values_centers}
\title{bin_values_centers}
\usage{
bin_values_centers(x, n_bins, xrng = range(x))
}
\arguments{
\item{x}{Values to assign to bins}

\item{n_bins}{Number of bins to assign values to}

\item{xrng}{Optional numeric of length 2. Defines the domain to define bins
for. Defaults to range(x).}
}
\value{
numeric of length n_bins containing centers of bins.
}
\description{
bin_values_centers
}
\examples{
bin_values_centers(0:10, 3)
}
