% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{calc_delta}
\alias{calc_delta}
\title{Title}
\usage{
calc_delta(tsne_dt, cell_a, cell_b, x_points, y_points = x_points,
  strategy = "normal")
}
\arguments{
\item{strategy}{}
}
\description{
Title
}
