% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctPlots.R
\name{ctPlotBinAggregates}
\alias{ctPlotBinAggregates}
\title{ctPlotBinAggregates}
\usage{
ctPlotBinAggregates(
  sts,
  feature_name = NULL,
  signal_name = NULL,
  xmin = -Inf,
  xmax = Inf,
  agg_FUN = max,
  xbins = 20,
  ybins = xbins
)
}
\arguments{
\item{ybins}{}
}
\description{
ctPlotBinAggregates
}
