% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctPlots.R
\name{ctPlotPointsAnnotation}
\alias{ctPlotPointsAnnotation}
\title{ctPlotPointsAnnotation}
\usage{
ctPlotPointsAnnotation(
  sts,
  meta_data = NULL,
  feature_name = NULL,
  signal_name = NULL,
  anno_var = NULL,
  anno_var_label = anno_var,
  bg_color = "gray60",
  return_data = FALSE
)
}
\arguments{
\item{sts}{A ssvTSNE object with ssvQC.prepSignal already called.}

\item{meta_data}{A data.frame containing "id" and anno_var. Will use cluster assignment if not set.}

\item{feature_name}{Feature name present in sts.  With default of NULL, first
feature name will be used.}

\item{signal_name}{Signal name present in sts.  With default of NULL, first
signal name will be used.}

\item{anno_var}{Variable to extract from meta_data for plotting.}

\item{anno_var_label}{Label to use for color legend. Default is anno_var.}

\item{bg_color}{Color to use for plot background. Default is "gray60".}

\item{return_data}{}
}
\value{
ggplot with color of t-SNE points determined by annotation in meta_data.
}
\description{
ctPlotPointsAnnotation
}
\examples{
data(sts.test)
ctPlotPointsAnnotation(sts)

#example using peak call
query_gr = sts$features_config$assessment_features$CTCF_features
meta_data = data.table(id = names(query_gr), MCF10CA1_CTCF_rep1 = query_gr$`MCF10CA1\nCTCF\nrep1`)
ctPlotPointsAnnotation(sts, meta_data = meta_data, anno_var = "MCF10CA1_CTCF_rep1")
}
