% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctPlots.R
\name{ctPlotSummaryProfiles}
\alias{ctPlotSummaryProfiles}
\title{ctPlotSummaryProfiles}
\usage{
ctPlotSummaryProfiles(
  sts,
  feature_name = NULL,
  signal_name = NULL,
  xbins = 20,
  ybins = xbins,
  profile_value = ssvQC:::val2var[sts$signal_config$plot_value],
  profile_value_label = ssvQC:::val2lab[sts$signal_config$plot_value],
  q_tall_values = NULL,
  q_wide_values = NULL,
  xrng = NULL,
  yrng = NULL,
  ylim = c(0, NA),
  ma_size = 2,
  n_splines = 10,
  p = NULL,
  N_floor = 0,
  N_ceiling = NULL,
  min_size = 0.3,
  return_data = FALSE
)
}
\arguments{
\item{return_data}{}
}
\description{
ctPlotSummaryProfiles
}
\examples{
data(sts.test)
ctPlotSummaryProfiles(sts)
}
