% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_tsne.R
\name{dt2mat}
\alias{dt2mat}
\title{dt2mat}
\usage{
dt2mat(prof_dt, marks)
}
\arguments{
\item{prof_dt}{tidy data.table to cast to a matrix}

\item{marks}{character vector of marks to cast}
}
\value{
a wide matrix with dimensions of nrows = (length(unique(id)) * number
  of cells) and ncols = (numbers of viewing bins * number of marks)
}
\description{
casts a tidy data.table, as returned from fetch_bam_dt or fetch_bw_dt to a
wide matrix compatible with Rtsne
}
\examples{
n_bins = 5
n_cells = 3
n_marks = 2
n_ids = 5
bins = (seq(n_bins)-1) / (n_bins-1)
cells = paste("cell", LETTERS[seq(n_cells)], sep = "_")
marks = rev(paste("mark", rev(letters)[seq(n_marks)], sep = "_"))
ids = paste("region", seq(n_ids), sep = "_")
dt = data.table(
    x = rep(bins, length(marks)*length(cells)),
    mark = rep(marks, length(ids), each = length(bins)),
    cell = rep(cells, length(ids), each= length(bins)*length(marks)),
    id = rep(ids, each = length(bins)*length(marks)*length(cells))
    )
dt$y = runif(nrow(dt))
mat = dt2mat(dt, marks)
mat
}
