% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_raster_summary.R
\name{plot_binned_aggregates}
\alias{plot_binned_aggregates}
\title{plot_binned_aggregates}
\usage{
plot_binned_aggregates(
  agg_dt,
  xbins = 50,
  ybins = xbins,
  xrng = NULL,
  yrng = NULL,
  val = "y",
  bxval = "tx",
  byval = "ty",
  facet_ = "wide_var",
  bin_met = mean,
  min_size = 1,
  return_data = FALSE
)
}
\arguments{
\item{return_data}{}
}
\description{
plot_binned_aggregates
}
\examples{
agg_dt = aggregate_signals(profile_dt)#[, .(y = mean(y)), .(id, tall_var, wide_var)]
agg_dt = merge(agg_dt, tsne_dt, by = c("id", "tall_var"))
#control resolution in x and y separately
plot_raster_summary(agg_dt, facet_ = c("tall_var", "wide_var"),
    xbins = 2, ybins = 3)
#filter out sparse regions
plot_raster_summary(agg_dt, facet_ = c("tall_var", "wide_var"),
    xbins = 12, min_size = 0, xrng =  c(-.5, .5))
}
