% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_path.R
\name{plot_outline}
\alias{plot_outline}
\title{plot_outline}
\usage{
plot_outline(
  tsne_dt,
  qtall_vars,
  id_to_plot = NULL,
  p = NULL,
  xrng = c(-0.5, 0.5),
  yrng = c(-0.5, 0.5),
  bg_color = "gray",
  line_color_mapping = "black",
  fill_color_mapping = "gray",
  label_type = c("text", "label", "none")[3],
  bg_points = 5000,
  arrow_FUN = NULL
)
}
\arguments{
\item{tsne_dt}{data.table with tsne info (tx, ty, id, and tall_var)}

\item{qtall_vars}{character vector of items in tsne_dt$tall_var}

\item{id_to_plot}{character vector of ids in tsne_dt$id}

\item{p}{exiting ggplot to add a layer onto.  Default of NULL creates a new
ggplot.}

\item{xrng}{numeric of length 2. passed to coord_cartesian xlim.  Not used if
p is specified.  Default is c(-.5, .5).}

\item{yrng}{numeric of length 2. passed to coord_cartesian ylim.  Not used if
p is specified.  Default is c(-.5, .5).}

\item{bg_color}{character. color to use for background points. Default is
"gray"}

\item{line_color_mapping}{character that is valid color. If less than length
of id_to_plot, recycled across specified id_to_plot.  Can be named vector
to completely specify id_to_plot.}

\item{fill_color_mapping}{character that is valid color. If less than length
of id_to_plot, recycled across specified id_to_plot.  Can be named vector
to completely specify id_to_plot.}

\item{label_type}{character.  one of c("text", "label", "none").  controls
how, if at all, plot objects are labelled.}

\item{bg_points}{number of points to plot in background.  if 0, only points
corresponding to id_to_plot are drawn.  if -1, no points at all are drawn.}

\item{arrow_FUN}{result of grid::arrow().  Default of NULL does not draw arrowheads.}
}
\value{
a ggplot
}
\description{
a ggplot where the position of id in every tall_var specified by qtall_vars is
connected in a polygon.  Allows the identification of both regions where ids
are stable/dynamic and individual ids that are particularly dynamic.
}
\details{
Good for looking at large numbers of ids with a modest number of tall_vars.
}
\examples{
data(tsne_dt)
plot_outline(tsne_dt, unique(tsne_dt$tall_var), unique(tsne_dt$id)[1:3])
plot_outline(tsne_dt, unique(tsne_dt$tall_var), unique(tsne_dt$id)[1:3],
    label_type = "none")
plot_outline(tsne_dt, unique(tsne_dt$tall_var), unique(tsne_dt$id)[1:3],
    label_type = "label")
}
