% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{plot_profiles_selected}
\alias{plot_profiles_selected}
\title{plot_profiles_selected}
\usage{
plot_profiles_selected(profile_dt, qtall_vars, id_to_plot,
  color_mapping = NULL)
}
\arguments{
\item{profile_dt}{a tidy data.table containing profile data.  variable names
are id, tall_var, wide_var, x, y.}

\item{qtall_vars}{character vector of tall_vars to show data for.}

\item{id_to_plot}{character vector of ids to show data for.}

\item{color_mapping}{color_mapping for wide_vars.}
}
\value{
a ggplot
}
\description{
plot_profiles_selected
}
\examples{
data(profile_dt)
plot_profiles_selected(profile_dt,
    qtall_vars = unique(profile_dt$tall_var)[1:2],
    id_to_plot = unique(profile_dt$id)[1:3]) +
    theme(panel.spacing = unit(.04, "npc"))
}
