% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{plot_profiles_selected}
\alias{plot_profiles_selected}
\title{Title}
\usage{
plot_profiles_selected(data_dt, qgr, qcells, tss_ids,
  color_mapping = c(H3K27me3 = "firebrick", H3K4me3 = "forestgreen"))
}
\arguments{
\item{tss_ids}{}
}
\description{
Title
}
