% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{plot_recentered_velocity}
\alias{plot_recentered_velocity}
\title{Title}
\usage{
plot_recentered_velocity(tsne_dt, cell_a, cell_b, n_points,
  points_as_background = FALSE, p = NULL, min_N = 0,
  id_to_plot = NULL, arrow_FUN = NULL, angle_as_color = FALSE)
}
\arguments{
\item{id_to_plot}{}
}
\description{
Title
}
