% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{plot_velocity_arrows}
\alias{plot_velocity_arrows}
\title{plot_velocity_arrows}
\usage{
plot_velocity_arrows(velocity_dt, p = NULL, id_to_plot = NULL,
  angle_as_color = TRUE, background_color = "lightgray")
}
\arguments{
\item{velocity_dt}{data.table with angle and distance info.}

\item{p}{an existing ggplot to overlay images onto.  Default of NULL starts a
new plot.}

\item{id_to_plot}{character.  ids in tsne_dt$id. Default of NULL causes all
ids to be used.}

\item{angle_as_color}{if TRUE, a rainbow like scale is applied to angle.}

\item{background_color}{character. color to use when plotting background
points.}
}
\value{
ggplot of arrows indicating position in tall_var line a to tall_var line b.
}
\description{
plot_velocity_arrows
}
\examples{
data("tsne_dt")
vel_dt = prep_velocity(tsne_dt,
    unique(tsne_dt$tall_var)[2], unique(tsne_dt$tall_var)[3])
plot_velocity_arrows(vel_dt)
}
