% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_velocity.R
\name{prep_velocity}
\alias{prep_velocity}
\title{prep_velocity}
\usage{
prep_velocity(tsne_dt, cell_a, cell_b, id_to_plot = NULL,
  max_plotted = 500, return_data = FALSE, delta.min = 0,
  delta.max = Inf, angle.min = 0, angle.max = 360,
  drop_backgroud = FALSE)
}
\arguments{
\item{drop_backgroud}{}
}
\description{
prep_velocity
}
