% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fetch.R
\name{stsFetchTsneInput}
\alias{stsFetchTsneInput}
\title{stsFetchTsneInput}
\usage{
stsFetchTsneInput(qdt, qgr, region_size = NULL, file_format = NULL,
  qwin = 50, qmet = "summary", cap_value = 20, agg_FUN = NULL,
  high_on_right = TRUE, bfc = BiocFileCache::BiocFileCache(),
  n_cores = getOption("mc.cores", 1), rname = digest::digest(list(qgr,
  qdt[, 1:3], qwin, qmet)), force_overwrite = FALSE)
}
\arguments{
\item{qdt}{data.table containing, file, cell, mark, and optionally
norm_factor}

\item{qgr}{GRanges of regions to fetch}

\item{qwin}{number of datapoints to use to view each region}

\item{qmet}{strategy to use in each region, summary or sample}

\item{cap_value}{maximum allowed value.  useful for outlier handling.}

\item{high_on_right}{if TRUE, profiles where highest signal is on the left
are flipped.}

\item{bfc}{BiocFileCache object to use to cache data}

\item{n_cores}{number of cores to use. Defaults to value of mc.cores or 1 if
mc.cores is not set.}

\item{rname}{rname to use with cache.  Default is a digest of arguments
used to fetch data.}

\item{force_overwrite}{if TRUE, any contents of cache are overwritten.}
}
\value{
a tidy data.table of profile data.
}
\description{
retrieve data from series of either bigwig or bam files to a tidy data.table
for later input to stsRunTsne
}
\details{
validates inputs prepares query_gr fetches either bam or bigwig based on
extension of first file transform and normalizes profiles after fetch
}
\examples{
data("query_gr")
bw_files = dir(system.file('extdata', package = "seqtsne"), pattern = ".bw$", full.names = TRUE)
cfg_dt = data.table(file = bw_files)
cfg_dt[, c("cell", "mark") := tstrsplit(basename(file), "_", keep = 1:2)]
cfg_dt = cfg_dt[cell \%in\% c("ESH1", "HUES48", "HUES64")]
cfg_dt[, norm_factor := ifelse(mark == "H3K4me3", .3, 1)]
profile_dt = stsFetchTsneInput(cfg_dt, query_gr)
profile_dt
}
