% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cluster.R
\name{stsPlotClusterProfiles}
\alias{stsPlotClusterProfiles}
\title{Title}
\usage{
stsPlotClusterProfiles(
  profile_dt,
  cluster_dt,
  cluster_ = "cluster_id",
  w = 0.05,
  h = 0.05
)
}
\arguments{
\item{h}{}
}
\description{
Title
}
