% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_image_plots.R
\name{stsPlotSummaryProfiles}
\alias{stsPlotSummaryProfiles}
\title{stsPlotSummaryProfiles}
\usage{
stsPlotSummaryProfiles(
  profile_dt,
  position_dt,
  x_points,
  y_points = x_points,
  q_tall_vars = NULL,
  q_wide_vars = NULL,
  xrng = range(position_dt$tx),
  yrng = range(position_dt$ty),
  plot_type = c("glyph", "raster")[1],
  rname = NULL,
  odir = file.path(tempdir(), rname),
  force_rewrite = FALSE,
  n_cores = getOption("mc.cores", 1),
  apply_norm = TRUE,
  ylim = c(0, 1),
  ma_size = 2,
  n_splines = 10,
  p = NULL,
  facet_byCell = FALSE,
  line_color_mapping = NULL,
  vertical_facet_mapping = NULL,
  N_floor = 0,
  N_ceiling = NULL,
  min_size = 0.3,
  return_data = FALSE
)
}
\arguments{
\item{profile_dt}{a tidy data.table for profile data as retrieved by
stsFetchTsneInput.  Expected variable names are id, tall_var, wide_var, x, and y.}

\item{position_dt}{a tidy data.table containing t-sne embedding.  Expected
variable names are tx, ty, id, and tall_var.}

\item{x_points}{numeric.  number of grid points to use in x dimension.}

\item{y_points}{numeric.  number of grid points to use in y dimension.
Defaults to same value as x_points.}

\item{q_tall_vars}{character vector of tall_vars to plot. Default of NULL plots all.}

\item{q_wide_vars}{character vector of wide_vars to plot.  Default of NULL plots all.}

\item{xrng}{view domain in x dimension, default is range of position_dt$tx.}

\item{yrng}{view domain in y dimension, default is range of position_dt$ty.}

\item{plot_type}{character, must be one of "glyph" or "raster".  raster
uses ggimage::geom_image to embed images where glyph uses GGally::glyphs}

\item{rname}{prefix for image files.  existing image files are used if
present.}

\item{odir}{output directory for image files.}

\item{force_rewrite}{if TRUE, images are overwritten even if they exist.}

\item{n_cores}{number of cores to use when writing images.  Default is value
of mc.cores option if set or 1.}

\item{apply_norm}{if TRUE, y values are trimmed to 95th percentile and
transformed ot domain of [0,1]. Default is TRUE.}

\item{ylim}{y-limits of regional summary plots.  Default of c(0, 1) is
compatible with apply_norm = TRUE.}

\item{ma_size}{moving average size when smoothing profiles.}

\item{n_splines}{number of points to interpolate with splines.}

\item{p}{an existing ggplot to overlay images onto.  Default of NULL starts a
new plot.}

\item{facet_byCell}{boolean. If TRUE, plots are facetted by tall_var.}

\item{line_color_mapping}{named vector of line color.  Names correspond to
values of profile_dt 'wide_var' variable and values are colors.}

\item{vertical_facet_mapping}{named vector of groups.  vertical_facet_mapping
names are wide_vars and order of first occurence determines vertical position
top to bottom.}

\item{N_floor}{The value of N to consider 0.  bins with N values <= N_floor
will be ignored.}

\item{N_ceiling}{The value of N to consider 1.  bins with N values >=
N_ceiling will have images drawn at full size.}

\item{min_size}{Numeric (0, 1]. The minimum size images to draw.  The default
of .3 draws images for all bins with N values >= 30% of the way from
N_floor to N_ceiling.}

\item{return_data}{if TRUE, data.table that would have been used to create
ggplot is returned instead.}
}
\value{
returns a ggplot containing images summarizing t-sne space at
  resolution determined by x_points and y_points with the space density
  mapped to size.
}
\description{
stsPlotSummaryProfiles
}
\examples{
data("profile_dt")
data("tsne_dt")
stsPlotSummaryProfiles(profile_dt, tsne_dt, x_points = 4)
}
